Restructured to use webpack hot reloading
