+++ .babelrc
@@ -1,3 +1,9 @@
 {
-  "presets": ["es2015", "stage-0", "react"]
+  "presets": [
+    "es2017",
+    "react"
+  ],
+  "plugins": [
+    "transform-runtime"
+  ]
 }
