+++ package.json
@@ -22,10 +22,11 @@
     "react-router": "^2.3.0",
     "react-tappable": "^0.8.1",
     "redux": "^3.4.0",
-    "superagent": "^1.8.3"
+    "superagent": "^1.8.3",
+    "webpack-hot-middleware": "^2.12.2"
   },
   "scripts": {
-    "start": "node server/index.js",
+    "start": "NODE_ENV=DEVELOPMENT webpack-dev-server --hot --inline --history-api-fallback --host 0.0.0.0 --port 3001 --content-base build",
     "build": "gulp build",
     "watch": "gulp"
   },
@@ -37,7 +38,9 @@
     "babel-core": "^6.0.20",
     "babel-eslint": "^4.1.3",
     "babel-loader": "^6.0.1",
+    "babel-plugin-transform-runtime": "^6.12.0",
     "babel-preset-es2015": "^6.6.0",
+    "babel-preset-es2017": "^1.6.1",
     "babel-preset-react": "^6.0.15",
     "babel-preset-stage-0": "^6.0.15",
     "eslint": "^1.10.3",
@@ -68,6 +71,7 @@
     "through2": "^2.0.0",
     "uid": "0.0.2",
     "webpack": "^1.13.1",
+    "webpack-dev-server": "^1.14.1",
     "yargs": "^3.30.0"
   }
 }
