+++ src/index.html
@@ -37,7 +37,7 @@
 
 		<!-- Container stores our app -->
 		<div id="app"></div>
-		<script src="/app.js"></script>
+		<script src="/main.js"></script>
 
 	  </body>
 </html>
